+++ ui/src/dashboards/containers/DashboardPage.js
@@ -1,11 +1,13 @@
 import React, {PropTypes} from 'react';
 import ReactTooltip from 'react-tooltip';
+import {Link} from 'react-router';
+import _ from 'lodash';
 
 import LayoutRenderer from 'shared/components/LayoutRenderer';
 import TimeRangeDropdown from '../../shared/components/TimeRangeDropdown';
 import timeRanges from 'hson!../../shared/data/timeRanges.hson';
 
-import {getDashboard} from '../apis';
+import {getDashboards} from '../apis';
 import {getSource} from 'shared/apis';
 
 const DashboardPage = React.createClass({
@@ -20,19 +22,24 @@ const DashboardPage = React.createClass({
     const fifteenMinutesIndex = 1;
 
     return {
+      dashboards: [],
       timeRange: timeRanges[fifteenMinutesIndex],
     };
   },
 
   componentDidMount() {
-    getDashboard(this.props.params.dashboardID).then((resp) => {
+    getDashboards().then((resp) => {
       getSource(this.props.params.sourceID).then(({data: source}) => {
         this.setState({
-          dashboard: resp.data,
+          dashboards: resp.data.dashboards,
           source,
         });
       });
     });
+  },
+
+  currentDashboard(dashboards, dashboardID) {
+    return _.find(dashboards, (d) => d.id.toString() === dashboardID);
   },
 
   renderDashboard(dashboard) {
@@ -74,8 +81,8 @@ const DashboardPage = React.createClass({
   },
 
   render() {
-    const {dashboard, timeRange} = this.state;
-    const dashboardName = dashboard ? dashboard.name : '';
+    const {dashboards, timeRange} = this.state;
+    const dashboard = this.currentDashboard(dashboards, this.props.params.dashboardID);
 
     return (
       <div className="page">
@@ -84,8 +91,20 @@ const DashboardPage = React.createClass({
             <div className="page-header__left">
               <div className="dropdown page-header-dropdown">
                 <button className="dropdown-toggle" type="button" data-toggle="dropdown">
-                  <span className="button-text">{dashboardName}</span>
+                  <span className="button-text">{dashboard ? dashboard.name : ''}</span>
+                  <span className="caret"></span>
                 </button>
+                <ul className="dropdown-menu" aria-labelledby="dropdownMenu1">
+                  {(dashboards).map((d, i) => {
+                    return (
+                      <li key={i}>
+                        <Link to={`/sources/${this.props.params.sourceID}/dashboards/${d.id}`} className="role-option">
+                          {d.name}
+                        </Link>
+                      </li>
+                    );
+                  })}
+                </ul>
               </div>
             </div>
             <div className="page-header__right">