+++ ui/src/data_explorer/containers/DataExplorer.js
@@ -4,41 +4,38 @@ import PanelBuilder from '../components/PanelBuilder';
 import Visualizations from '../components/Visualizations';
 import Header from '../containers/Header';
 import ResizeContainer from 'shared/components/ResizeContainer';
-import {FETCHING} from '../reducers/explorers';
+
 import {
   setTimeRange as setTimeRangeAction,
-  createExploration as createExplorationAction,
-  chooseExploration as chooseExplorationAction,
-  deleteExplorer as deleteExplorerAction,
-  editExplorer as editExplorerAction,
 } from '../actions/view';
+
+const {
+  func,
+  shape,
+  string,
+} = PropTypes;
 
 const DataExplorer = React.createClass({
   propTypes: {
-    source: PropTypes.shape({
-      links: PropTypes.shape({
-        proxy: PropTypes.string.isRequired,
-        self: PropTypes.string.isRequired,
+    source: shape({
+      links: shape({
+        proxy: string.isRequired,
+        self: string.isRequired,
       }).isRequired,
     }).isRequired,
-    explorers: PropTypes.shape({}).isRequired,
-    explorerID: PropTypes.string,
-    timeRange: PropTypes.shape({
-      upper: PropTypes.string,
-      lower: PropTypes.string,
+    timeRange: shape({
+      upper: string,
+      lower: string,
     }).isRequired,
-    setTimeRange: PropTypes.func.isRequired,
-    createExploration: PropTypes.func.isRequired,
-    chooseExploration: PropTypes.func.isRequired,
-    deleteExplorer: PropTypes.func.isRequired,
-    editExplorer: PropTypes.func.isRequired,
+    activePanel: string,
+    setTimeRange: func.isRequired,
   },
 
   childContextTypes: {
-    source: PropTypes.shape({
-      links: PropTypes.shape({
-        proxy: PropTypes.string.isRequired,
-        self: PropTypes.string.isRequired,
+    source: shape({
+      links: shape({
+        proxy: string.isRequired,
+        self: string.isRequired,
       }).isRequired,
     }).isRequired,
   },
@@ -49,13 +46,8 @@ const DataExplorer = React.createClass({
 
   getInitialState() {
     return {
-      activePanelID: null,
       activeQueryID: null,
     };
-  },
-
-  handleSetActivePanel(id) {
-    this.setState({activePanelID: id});
   },
 
   handleSetActiveQuery(id) {
@@ -63,32 +55,24 @@ const DataExplorer = React.createClass({
   },
 
   render() {
-    const {timeRange, explorers, explorerID, setTimeRange, createExploration, chooseExploration, deleteExplorer, editExplorer} = this.props;
-
-    if (explorers === FETCHING || !explorerID) {
-      // TODO: page-wide spinner
-      return null;
-    }
+    const {timeRange, setTimeRange, activePanel} = this.props;
 
     return (
       <div className="data-explorer">
         <Header
-          actions={{setTimeRange, createExploration, chooseExploration, deleteExplorer, editExplorer}}
-          explorers={explorers}
+          actions={{setTimeRange}}
           timeRange={timeRange}
-          explorerID={explorerID}
         />
         <ResizeContainer>
           <PanelBuilder
             timeRange={timeRange}
-            activePanelID={this.state.activePanelID}
+            activePanelID={activePanel}
             activeQueryID={this.state.activeQueryID}
             setActiveQuery={this.handleSetActiveQuery}
-            setActivePanel={this.handleSetActivePanel}
           />
           <Visualizations
             timeRange={timeRange}
-            activePanelID={this.state.activePanelID}
+            activePanelID={activePanel}
             activeQueryID={this.state.activeQueryID}
           />
         </ResizeContainer>
@@ -98,16 +82,14 @@ const DataExplorer = React.createClass({
 });
 
 function mapStateToProps(state) {
+  const {timeRange, dataExplorerUI} = state;
+
   return {
-    timeRange: state.timeRange,
-    explorers: state.explorers,
+    timeRange,
+    activePanel: dataExplorerUI.activePanel,
   };
 }
 
 export default connect(mapStateToProps, {
   setTimeRange: setTimeRangeAction,
-  createExploration: createExplorationAction,
-  chooseExploration: chooseExplorationAction,
-  deleteExplorer: deleteExplorerAction,
-  editExplorer: editExplorerAction,
 })(DataExplorer);