+++ ui/src/data_explorer/reducers/dataExplorerUI.js
@@ -1,0 +1,11 @@
+export default function dataExplorerUI(state = {}, action) {
+  switch (action.type) {
+    case 'ACTIVATE_PANEL':
+    case 'CREATE_PANEL': {
+      const {panelID} = action.payload;
+      return {...state, activePanel: panelID};
+    }
+  }
+
+  return state;
+}