+++ ui/src/data_explorer/reducers/index.js
@@ -1,13 +1,11 @@
 import queryConfigs from './queryConfigs';
 import panels from './panels';
 import timeRange from './timeRange';
-import explorers from './explorers';
-import activeExplorer from './activeExplorer';
+import dataExplorerUI from './dataExplorerUI';
 
 export {
   queryConfigs,
   panels,
   timeRange,
-  explorers,
-  activeExplorer,
+  dataExplorerUI,
 };