+++ ui/src/store/configureStore.js
@@ -2,14 +2,14 @@ import {createStore, applyMiddleware, compose} from 'redux';
 import {combineReducers} from 'redux';
 import thunkMiddleware from 'redux-thunk';
 import makeQueryExecuter from 'src/shared/middleware/queryExecuter';
-import * as chronografReducers from 'src/data_explorer/reducers';
+import * as dataExplorerReducers from 'src/data_explorer/reducers';
 import * as sharedReducers from 'src/shared/reducers';
 import rulesReducer from 'src/kapacitor/reducers/rules';
 import persistStateEnhancer from './persistStateEnhancer';
 
 const rootReducer = combineReducers({
   ...sharedReducers,
-  ...chronografReducers,
+  ...dataExplorerReducers,
   rules: rulesReducer,
 });